# Introducing lookup3 (UTF-8 validation).