This slightly improves stage 1 (UTF-8 validation).

On x64 with GNU GCC 9 


|filename | ref. instructions/byte |candidate instructions/byte | difference |
|:--------| ----------:| --------------:| ----------:|
|apache_builds.json|2.310|2.302|-0.3%|
|canada.json|2.596|2.588|-0.3%|
|citm_catalog.json|2.176|2.168|-0.4%|
|github_events.json|2.149|2.138|-0.5%|
|google_maps_api_compact_response.json|3.113|3.110|-0.1%|
|google_maps_api_response.json|2.279|2.271|-0.3%|
|gsoc-2018.json|1.685|1.666|-1.1%|
|instruments.json|2.370|2.362|-0.3%|
|marine_ik.json|2.959|2.952|-0.2%|
|mesh.json|2.852|2.846|-0.2%|
|mesh.pretty.json|2.285|2.277|-0.3%|
|numbers.json|2.387|2.380|-0.3%|
|random.json|3.424|3.391|-1.0%|
|repeat.json|3.105|3.066|-1.2%|
|tree-pretty.json|2.331|2.323|-0.3%|
|twitter.json|2.453|2.434|-0.8%|
|twitter_api_compact_response.json|2.528|2.516|-0.5%|
|twitter_api_response.json|2.325|2.314|-0.5%|
|twitter_timeline.json|2.443|2.434|-0.4%|
|twitterescaped.json|2.189|2.177|-0.5%|
|update-center.json|2.348|2.339|-0.4%|

Variation: -1.2% -- -0.1%


And on a rockpro64 (Cortex-A72):

|filename | ref. instructions/byte |candidate instructions/byte | difference |
|:--------| ----------:| --------------:| ----------:|
|apache_builds.json|2.967|2.966|-0.0%|
|canada.json|3.301|3.301|-0.0%|
|citm_catalog.json|2.811|2.810|-0.0%|
|github_events.json|2.787|2.784|-0.1%|
|google_maps_api_compact_response.json|3.857|3.853|-0.1%|
|google_maps_api_response.json|2.935|2.929|-0.2%|
|gsoc-2018.json|2.266|2.266|-0.0%|
|instruments.json|3.037|3.036|-0.0%|
|marine_ik.json|3.704|3.706|0.1%|
|mesh.json|3.564|3.569|0.1%|
|mesh.pretty.json|2.938|2.938|-0.0%|
|numbers.json|3.058|3.057|-0.0%|
|random.json|4.337|4.206|-3.0%|
|repeat.json|3.991|3.816|-4.4%|
|tree-pretty.json|2.994|2.990|-0.1%|
|twitter.json|3.164|3.117|-1.5%|
|twitter_api_compact_response.json|3.237|3.210|-0.8%|
|twitter_api_response.json|2.991|2.974|-0.6%|
|twitter_timeline.json|3.124|3.121|-0.1%|
|twitterescaped.json|2.832|2.833|0.0%|
|update-center.json|3.015|3.015|-0.0%|

Variation: -4.4% -- 0.1%



This is not a large gain but from-principle analysis suggests that the core part of the lookup3 algorithm is cheaper than the core part of the lookup2 algorithm. What makes it a bit harder to to assess is that this core is mixed in as part of larger complex kernels. (Diminishing returns and all that.)

The effect is likely too small to measure in terms of GB/s, but there should be some benefit.