@lemire 2020-06-16
@jkeiser Marking tentatively for 0.4.