@jkeiser 2020-06-16
I'm pretty sure this will improve throughput too ... but did you have numbers on that? I've made more than a few PRs recently that reduce instructions but still hurt throughput :)