@jkeiser 2020-06-16
Here are some skylake numbers for reference. It is absolutely, unabashedly positive :)

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| twitter                             |  45.4 |  44.2 |    2% | 156.8 | 155.7 |    0% |   1176 |   1025 |   12% |      0 |      0 |    0% |  29242 |  29257 |    0% |
| repeat                              |  56.0 |  54.6 |    2% | 197.8 | 195.3 |    1% |      2 |      1 |   50% |      0 |      0 |    0% |      0 |      4 |    0% |
| random                              |  61.2 |  59.7 |    2% | 219.1 | 217.0 |    0% |    783 |    679 |   13% |      0 |      0 |    0% |  29826 |  29757 |    0% |
| update-center                       |  46.3 |  45.7 |    1% | 150.2 | 149.7 |    0% |   2128 |   2055 |    3% |      0 |      0 |    0% |  27120 |  26919 |    0% |
| twitter_api_response                |  41.6 |  41.0 |    1% | 148.1 | 147.5 |    0% |      2 |      3 |  -50% |      0 |      0 |    0% |     88 |    118 |  -34% |
| tree-pretty                         |  40.4 |  40.0 |    1% | 149.1 | 148.6 |    0% |      2 |      2 |    0% |      0 |      0 |    0% |    422 |    240 |   43% |
| instruments                         |  40.9 |  40.4 |    1% | 151.6 | 151.1 |    0% |     37 |     19 |   48% |      0 |      0 |    0% |  10991 |  11309 |   -2% |
| numbers                             |  41.2 |  40.8 |    1% | 152.7 | 152.2 |    0% |     34 |     17 |   50% |      0 |      0 |    0% |   7451 |   6568 |   11% |
| canada                              |  48.8 |  48.3 |    1% | 166.1 | 165.6 |    0% |   5219 |   5184 |    0% |  13700 |  11139 |   18% | 122613 | 122680 |    0% |
| twitter_api_compact_response        |  45.8 |  45.4 |    0% | 160.8 | 160.1 |    0% |      2 |      2 |    0% |      0 |      0 |    0% |     24 |      5 |   79% |
| apache_builds                       |  40.0 |  39.6 |    0% | 147.8 | 147.3 |    0% |     13 |      4 |   69% |      0 |      0 |    0% |   4996 |   5242 |   -4% |
| mesh                                |  49.3 |  48.9 |    0% | 182.7 | 182.2 |    0% |    618 |    541 |   12% |      0 |      0 |    0% |  45922 |  46145 |    0% |
| citm_catalog                        |  37.3 |  37.0 |    0% | 139.2 | 138.7 |    0% |    144 |    144 |    0% |    427 |     97 |   77% |  76970 |  77282 |    0% |
| google_maps_api_response            |  39.8 |  39.5 |    0% | 145.7 | 145.2 |    0% |      3 |      2 |   33% |      0 |      0 |    0% |     93 |     52 |   44% |
| google_maps_api_compact_response    |  54.5 |  54.1 |    0% | 198.8 | 198.3 |    0% |      3 |      2 |   33% |      0 |      0 |    0% |    276 |     50 |   81% |
| marine_ik                           |  57.6 |  57.2 |    0% | 189.4 | 188.9 |    0% |   8304 |   8387 |    0% |  77945 |  64080 |   17% | 191776 | 191467 |    0% |
| mesh.pretty                         |  39.0 |  38.8 |    0% | 146.2 | 145.7 |    0% |     36 |     30 |   16% |    142 |    129 |    9% |  74392 |  74814 |    0% |
| twitterescaped                      |  42.5 |  42.3 |    0% | 139.8 | 139.3 |    0% |   1704 |   1747 |   -2% |      0 |      0 |    0% |  26815 |  26829 |    0% |
| github_events                       |  37.4 |  37.2 |    0% | 137.3 | 136.8 |    0% |      5 |      7 |  -40% |      0 |      0 |    0% |   1193 |   1681 |  -40% |
| twitter_timeline                    |  42.4 |  42.3 |    0% | 156.2 | 155.7 |    0% |      3 |      5 |  -66% |      0 |      0 |    0% |    778 |   1010 |  -29% |
| gsoc-2018                           |  34.4 |  34.6 |    0% | 107.1 | 106.6 |    0% |   6136 |   6290 |   -2% |  30096 |  26251 |   12% | 122046 | 122450 |    0% |