@lemire 2020-06-17
@jkeiser Do you approve the PR? To be clear, this sets lookup3 as the new default. I have high confidence that it is correct.