@jkeiser 2020-06-17
@lemire I'm looking it over :) Sorry, I was a bit out of the loop yesterday.