@lemire 2020-06-17
Merged.